+++ frontend/src/components/StableCoin/Analytics.vue
@@ -18,7 +18,7 @@
 						:is-loading="isLoading"
 						:transfer-summary="pltTransferMetricsDataRef"
 						:decimals="
-							pltTransferMetricsDataRef?.pltTransferMetrics.decimal || 0
+							pltTransferMetricsDataRef?.pltTransferMetricsByTokenId.decimal || 0
 						"
 					/>
 				</CarouselSlide>