+++ frontend/src/queries/usePltTransferMetricsQuery.ts
@@ -2,20 +2,20 @@ import { useQuery, gql } from '@urql/vue'
 import type { Ref } from 'vue'
 import type {
 	MetricsPeriod,
-	PltTransferMetrics,
-	PltMetrics,
+	PltTransferMetricsByTokenId,
+	GlobalPltMetrics,
 } from '~/types/generated'
 
 export type PltMetricsQueryResponse = {
-	pltMetrics: PltMetrics
+	globalPltMetrics: GlobalPltMetrics
 }
 
 const PltMetricsQuery = gql<PltMetricsQueryResponse>`
-	query PltMetrics($period: MetricsPeriod!) {
-		pltMetrics(period: $period) {
-			transactionCount
+	query GlobalPltMetrics($period: MetricsPeriod!) {
+		globalPltMetrics(period: $period) {
+			eventCount
 			transferVolume
-			uniqueAccounts
+			
 		}
 	}
 `
@@ -31,12 +31,12 @@ export const usePltMetricsQuery = (period: Ref<MetricsPeriod>) => {
 }
 
 export type PltTransferMetricsQueryResponse = {
-	pltTransferMetrics: PltTransferMetrics
+	pltTransferMetricsByTokenId: PltTransferMetricsByTokenId
 }
 
 const PltTransferMetricsQueryByTokenId = gql<PltTransferMetricsQueryResponse>`
-	query PltTransferMetrics($period: MetricsPeriod!, $tokenId: String!) {
-		pltTransferMetrics(period: $period, tokenId: $tokenId) {
+	query PltTransferMetricsByTokenId($period: MetricsPeriod!, $tokenId: String!) {
+		pltTransferMetricsByTokenId(period: $period, tokenId: $tokenId) {
 			transferCount
 			transferVolume
 			decimal